add tests for Six command processor class
I've had to refactor the surround() function so it returns the
defined class. All this annoying dancing is forced on us because
we need a way to detect whether Six is available before loading
the plugin. I can't think of a cleaner way of doing this.

Also, surround() returns the defined class because we need to
access it in tests.